refactor: one home for the wic: vocabulary, one home for the generators
The set of keys a `wic:` block admits lived only in a dict literal inside
the validating schema. Nothing linked it to the language, so `lang_version`
reached the reference, the resolver and the compiler while the validator
still rejected the document before compilation began.

`Grammar.SIDECAR_KEYS` now names that vocabulary where the rest of the
language is named, and a test requires the validator to admit exactly those
keys. The shapes stay with the validator; only the vocabulary is shared.

The generators move to `tests/core/strategies.py`. Three suites quantify
over the same language, and a second generator beside the first is how
outputs and nested sidecars once became invisible to every property fed by
one of them. Extracted unchanged.

Also removed: a `scalars` strategy with no users, and a tag-to-node test
whose four cases the surface-equivalence property already covers. The one
case it covered alone — a bare name is an unresolved name — is kept, since
that form has no tagged/desugared pair for the property to reach.
